+++ ci/docker_build_images.ps1
@@ -66,17 +66,17 @@ function Start-DockerBuild  {
 }
 
 $Builds = [DockerBuild[]] @(
-    , [DockerBuild]::new('package-unity-tester', 'v1-2018.4.14f1').
-        WithExtraArg('--build-arg', 'UNITY_VERSION=2018.4.14f1')
+#   , [DockerBuild]::new('package-unity-tester', 'v1-2018.4.14f1').
+#       WithExtraArg('--build-arg', 'UNITY_VERSION=2018.4.14f1')
 
-    , [DockerBuild]::new('package-unity-tester', 'v1-2019.2.11f1').
-        WithExtraArg('--build-arg', 'UNITY_VERSION=2019.2.11f1')
+#   , [DockerBuild]::new('package-unity-tester', 'v1-2019.2.11f1').
+#       WithExtraArg('--build-arg', 'UNITY_VERSION=2019.2.11f1')
 
-    , [DockerBuild]::new('package-builder', 'v2')
+    , [DockerBuild]::new('package-builder', 'v3')
 
-    , [DockerBuild]::new('package-deploy-npm', 'v3')
+#   , [DockerBuild]::new('package-deploy-npm', 'v3')
 
-    , [DockerBuild]::new('package-deploy-github', 'v6')
+#   , [DockerBuild]::new('package-deploy-github', 'v6')
 )
 
 $Builds | Start-DockerBuild