+++ ci/scripts/get_framework_from_build.sh
@@ -5,10 +5,10 @@ set -o nounset
 set -o errexit
 set -o pipefail
 
-build="${1?"Build name required. Possible values: 'Standalone', 'AOT', 'Portable', 'Editor', 'Tests'."}"
+build="${1?"Build name required. Possible values: 'Standalone', 'AOT', 'Editor', 'Tests'."}"
 
 error() {
-    >&2 echo "$0: $@"
+    >&2 echo "$0: ERROR: $@"
     exit 1
 }
 
@@ -20,7 +20,8 @@ AOT)
     framework="netstandard2.0"
     ;;
 Portable)
-    framework="portable-net45+win8+wpa81+wp8"
+    error "Portable Json .NET build has been deprecated since 13.0.1."
+    exit 1
     ;;
 Editor)
     framework="netstandard2.0"
@@ -30,7 +31,7 @@ Tests)
     ;;
 *)
     error "Invalid build name.
-    Possible values: 'Standalone', 'AOT', 'Portable', 'Editor', 'Tests'."
+    Possible values: 'Standalone', 'AOT', 'Editor', 'Tests'."
     ;;
 esac